Temporarily remove softc typechecks and simplify device_get_softc
This commit allows using device_get_softc with pointers in the Attaching state.
In what is expected to be the most common use case the device_t pointer in a
softc (if any) will be in this state. There really is no reason to pass an
sc->sc_dev to device_get_softc so this code path has a small runtime check
unlike the fast path of using the device_t passed directly to the interface
method. This cost can be avoided by replacing sc->sc_dev after the softc on the
heap has been initialized.