Build rust sysroot from compiler source
Previously libcore.rlib and libcompiler_builtins.rlib came from the system rust
installation for the arch for the given kernel target. This is quite inflexible
when rustup is not available (e.g. FreeBSD aarch64) or the recommended way to
install rust. Also the prebuilt rlibs may need to be built with flags matching
the rest of the kernel objects in some cases (e.g. BTI and fixed-x18 on aarch64)
so per-config prebuilts would be needed.

This commit removes the need for using these prebuilts by building a barebones
rust sysroot from the compiler source. The only part that is really needed is
libcore.rlib but omitting libcompiler_builtins.rlib triggers a compile error so
a fake rlib with panicking stubs is provided for this. For libcore the intended
build process is to clone the rust-lang/rust repo into src/sys/rust/compiler
and update its stdarch submodule. The compiler does not actually need to be
built to build libcore.rlib though this does mean that it requires using a
nightly version. For now the plan is to keep the KPI crate and drivers building
on stable (using cargo test in CI to ensure this for the KPI crate) and
eventually move to stable once it becomes possible to build libcore without a
nightly rustc.