kpi: Replace device_{init,get}_softc with interface prefix args
This commit adds the softc as the first argument to interface trait methods.
This allows removing the device_init_softc and device_get_softc macros/functions
which greatly simplifies the checks that need to be done by the KPI crate's glue
code. It also removes the need to grab a refcount on each call to an interface
method. It also treats device_probe and device_attach as special cases since the
former doesn't have an associated softc at that point and the latter starts off
with the softc in an uninitialized state. To soundly allow initializing the
softc in device_attach with minimal chance of invoking UB this commit also adds
an Uninit type which will only exposes a single method for initializing the
softc without making the user take into account the details of `RefCounted<T>`.
The init method then returns a `&RefCounted<Softc>` to allow using it to create
`Ptr<T>` and to project it.