kpi/intr: Loosen static bound on intr_irqsrc out pointers
The pic interface has two methods which require setting intr_irqsrc out
pointers. Previously the PicIf trait forced these out pointers to have a static
lifetime. This commit uses the fact that the softc pointer is now an argument to
the methods to loosen this bound. Now the out pointers need to have the same
lifetime as the softc pointer (or 'static). While interrupt controllers still
have to leak their softc (due to intr_pic_claim_root and the isrc/ipi register
functions) this would make it easier to remove that requirement in the future if
it becomes possible. It also removes an atomic access on the refcount when
setting up each intr_irqsrc (because it avoids the need to call leak_ref).