kpi/tests: Remove CDriverFns trait
This commit replaces the CDriverFns used in test infrastructure with code that
follows the kobj_class structure to get each interface method. This reduces the
amount of boilerplate needed to add a new test driver and makes the userspace
tests operation more similar to what the kernel does as well.