kpi/bindings.h: Start using __has_include
This header is mainly for use by bindgen (ignoring when it's compiled for macro
instantiations which could be split out into a separate source file). This means
that it will be preprocessed by libclang (because bindgen does not support
anything else). Since it will only be run through libclang this commit replaces
conditional compilation based on arch with #if __has_include. Arch-based
conditional would not work for KERNCONFs which don't include support for the
headers they're trying to include. It has only worked until now since I've only
been routinely testing GENERIC/MINIMAL on amd64 and APPLE/VIRT on arm64. The
ideal fix would be for the build system to generate preprocessor variables from
makefile flags (MK_FDT leads to #define WITH_MK_FDT 1 or similar) but for now
this is a more maintainable fix as I start to support more KERNCONFs.