ci: Build demos and run tests on dev branch
The dev branch contains WIP dev work intended for main. It should generally stay
rebased on main for ease of merging, but there are no guarantees provided about
its build status. The test branch is intended for running just userspace unit
tests in CI. Do not expect it to remain rebased on main or even exist in
general.