kern.rust.mk: Remove test_bindings.rs from list of KPI files
The libkpi.rlib target should not depend on the bindings file at the root of the
crate (previously bindings.rs, now test_bindings.rs) because this is intended
only for userspace unit tests. Previously it was a dependency though not linked
into libkpi.rlib (because the bindings module checked the std feature to
determine which binding.rs file to use). This commit removes that incorrect
dependency to fix the broken build caused by renaming the file.