kpi/tests: Add feature for arch-specific unit tests
Both features run the test on the host (and create test binaries for the host
arch), but the functions supported by aarch64 and x86_64 differ a decent bit so
this commit adds bindings from amd64/KERNCONF=MINIMAL and arm64/KERCONF=VIRT. It
also changes the KPI crate type to rlib to allow writing unit tests in crates
that depend on this one. For now x86_64 tests are disabled since all the tests
using the driver! macro use ofw bus to probe for devices.