kpi/collections: Make ListBuffer Drop impl panic
Keeping the ListBuffer API simple requires forcing some level of manual memory
management on the user. If ListBuffer's Drop impl frees the buffer this can lead
to errors that are hard to debug (e.g. virtio devices reading deadc0de in
descriptor buffers). If the ListBuffer's Drop impl mem::forget's the buffer that
would lead to potential memory leaks that are hard to diagnose since Drop is
implicit in the source. The solution for now is to make ListBuffer's Drop impl
panic and require calling one of its methods to either get the buffer once the
sglist has been reset or leak the buffer if it's a negligible amount of memory.
This does not rely on the Drop impl running to maintain memory-safety since
calling mem::forget is equivalent to calling list_buffer.leak_buffer() with the
addition that the sglist refcount held by the ListBuffer is also leaked.