kpi: Support rustc 1.85.0
rustc 1.85 is the oldest version that supports edition 2024, but some KPI code
(mostly tests) used APIs that had been stabilized in newer versions. This commit
reworks this code to work with rustc 1.85 and updates the CI workflow to use
this as the oldest version for userspace tests. This ensures the widest possible
range of rustc versions supporting edition 2024 can be used.