kpi/arc: Add Arc::take_refcount and fix tests
The unmangled device_free_softc used in tests assumes the first field is a
device_t so this commit fixes test drivers which did not have repr(C) on the
softc. This is necessary to test out the new Arc::take_refcount method and
ArcRefCount types which allow releasing and potentially freeing an Arc without
knowning its type. This will be used by various functions which take ownership
of an Arc (e.g. config_hook, callout, etc.) but don't want to require specifying
the Arc's type in the wrapper doing the refcount_release.