kpi/ffi: Switch from refcounted softc to DevRef
This commit switches softc refcounting for a new DevRef type. While I'm still
figuring out the invariants this type enforces, the gist is that it's a
reference that is known to live as long as a device_t's softc. It's currently a
thin pointer to make interop with C callbacks easier. I've considered making it
a fat pointer (by adding the device_t) to allow pointer projection but it may be
better to keep it as a thin pointer. That would also allow keeping the invariant
that the pointee is always the start of the malloc'ed softc mem which would make
it easy to use bindings::device_get_softc(dev) to determine if a given DevRef is
paired with a device_t. Currently many of the callback KPIs do not check this
but they really should. There are some KPIs where there is no associated
device_t. These need to be looked at on a one by one basis, but for most of them
I'm leaning towards just requiring an Arc.